agent/requirements.txt
livekit-agents
livekit-plugins-bey
livekit-plugins-openai
livekit-plugins-silero
livekit-plugins-deepgram
livekit-plugins-cartesia
livekit-plugins-turn-detector
python-dotenv
livekit-api
uvicorn
fastapi
pydantic
requests
supabase
